pyproject.toml
[project]
name = "careerpython"
version = "0.1.0"
description = ""
authors = [{name = "juanmacias", email = "extjmv@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "alembic==1.13.2",
    "argon2-cffi==23.1.0",
    "bcrypt==4.3.0",
    "dependency-injector==4.48.1",
    "fastapi==0.111.0",
    "httptools==0.6.4",
    "passlib==1.7.4",
    "pipdeptree==2.28.0",
    "psycopg2-binary==2.9.10",
    "pydantic-settings==2.3.4",
    "pytest==8.3.2",
    "pytest-asyncio==0.24.0",
    "pytest-cov>=5.0.0",
    "httpx==0.27.0",
    "python-jose==3.3.0",
    "pyyaml==6.0.2",
    "ulid-py==1.1.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "uvicorn>=0.35.0,<0.36.0",
    "pypdf2==3.0.1",
    "markdown==3.7",
    "beautifulsoup4==4.12.3",
    "requests==2.32.3",
    "reportlab==4.2.5",
    "weasyprint==62.3",
    "factory-boy==3.3.1",
    "faker==30.8.2",
    "mypy",
    "dramatiq[redis]==1.15.0",
    "redis==5.0.1",
    "flake8>=7.3.0",
    "autopep8>=2.3.2",
    "autoflake>=2.3.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.17.1,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
